Switched price value to `BigDecimal`

+++ app/src/main/java/dev/supasintatiyanupanwong/apps/android/bpi/prices/domain/models/PriceInfo.kt
@@ -1,8 +1,9 @@
 package dev.supasintatiyanupanwong.apps.android.bpi.prices.domain.models
 
+import java.math.BigDecimal
 import java.util.Currency
 
 data class PriceInfo(
     val currency: Currency,
-    val value: Double
+    val value: BigDecimal
 )
